refactor(demos): migrate wave 8 (on-change, validate-on-modes) to the shared style system
Move both flat to folder with zero bespoke residue, and grow the shared grid
and header vocab so a third demo can shed its residue too:

- layout: add a `split3` modifier, the three-column sibling of `split`.
  validate-on-modes uses it for its three side-by-side mode cards, and
  schema-defaults drops its hand-rolled `repeat(3, 1fr)` media query onto it
  (keeping only its genuinely bespoke `.blank` marker).
- header: add a `.spread` utility (the same space-between row as a standalone
  class) so on-change's field caption can carry a trailing status pill without
  reinventing the row.

on-change also collapses .toggle onto row compact and reuses the status pill
fragment as-is. validate-on-modes maps its mode cards onto card + a nested
section heading, and its raw-result box onto the now-block banner (neutral,
tinting danger on error) instead of a one-off mono box.
